scaling-palm-tree/backend/requirements.txt
fastapi
uvicorn
motor
langchain
langchain-google-genai
pandas
python-dotenv
pydantic
